requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
plotly>=5.18.0
openai>=1.12.0
fpdf2>=2.7.7
python-dotenv>=1.0.0
